+++ CHANGELOG.md
@@ -2,12 +2,14 @@
 
 ## master (unreleased)
 
+## 4.0.0 (12/12/2014)
+
 ### Changes
 
 * Removed `inferior-lisp` integration in favor of `inf-clojure`.
 * Indented the body of `cond` with 2 spaces.
 * Removed special indentation settings for `defstruct`, `struct-map` and `assoc`.
-* Added special indentation settings for `->`, `->>`, `cond->` and `cond->>`.
+* Added special indentation settings for `->`, `->>`, `cond->`, `cond->>`, `some->` and `some->>`.
 
 ## 3.0.1 (24/11/2014)